Move additional feature SCSS modules to CSS
