--- components/about/About.module.scss
@@ -1,145 +1,0 @@
-@use "../../styles/variables.scss";
-
-.calendarTable {
-  border: 1px solid variables.$almost-very-dark-grey !important;
-  tr td {
-    padding: 10px;
-    vertical-align: middle;
-  }
-
-  td {
-    line-height: 40px;
-  }
-
-  tr:nth-child(odd) {
-    background-color: variables.$almost-very-dark-grey;
-  }
-}
-
-.activeSection {
-  border: 1px solid variables.$off-white !important;
-}
-
-.cookiePolicyTable {
-  th {
-    padding-top: 25px;
-    padding-bottom: 25px;
-  }
-  td {
-    border: 1px solid white !important;
-    width: fit-content;
-    padding: 10px;
-  }
-}
-
-.lastUpdateText {
-  color: variables.$font-color-h !important;
-}
-
-.htmlContainer {
-  h2,
-  h3,
-  h4,
-  h5 {
-    font-family: variables.$main-font !important;
-    float: none;
-    color: variables.$font-color;
-  }
-}
-
-.faqContainer {
-  color: #cecfd4;
-
-  h2,
-  h3 {
-    font-family: variables.$main-font !important;
-    float: none;
-    color: variables.$font-color;
-  }
-
-  h2 {
-    margin-top: 0;
-    margin-bottom: 1.5rem;
-  }
-
-  h3 {
-    margin-bottom: 0.75rem;
-    font-size: 1.25rem;
-    line-height: 1.35;
-  }
-
-  p,
-  li,
-  dd {
-    line-height: 1.65;
-  }
-
-  p,
-  ul,
-  ol,
-  dl {
-    margin-bottom: 1rem;
-  }
-
-  ul,
-  ol {
-    padding-left: 1.5rem;
-  }
-
-  dt {
-    margin-top: 1rem;
-    margin-bottom: 0.35rem;
-    color: #efeff1;
-    font-weight: 700;
-  }
-
-  dd {
-    margin-left: 0;
-  }
-
-  a {
-    color: #528bff;
-    font-weight: 600;
-    overflow-wrap: anywhere;
-    text-decoration-color: rgb(82 139 255 / 45%);
-    text-underline-offset: 0.18em;
-
-    &:hover,
-    &:focus-visible {
-      color: #84adff;
-      text-decoration-color: currentColor;
-    }
-  }
-}
-
-.faqList {
-  padding-left: 1.75rem;
-
-  > li {
-    margin-bottom: 2rem;
-    padding-left: 0.25rem;
-  }
-
-  > li::marker {
-    font-size: 1.2rem;
-    font-weight: 700;
-  }
-}
-
-.faqImage {
-  display: block;
-  width: 100%;
-  height: auto;
-  margin-top: 1.25rem;
-  margin-bottom: 1.5rem;
-  border-radius: 6px;
-}
-
-.faqBonus,
-.walletSection {
-  margin-top: 2.5rem;
-}
-
-.walletSection {
-  scroll-margin-top: 5rem;
-}
