+++ components/about/AboutCopyright.tsx
@@ -3,7 +3,7 @@ import {
   AboutContainer as Container,
   AboutRow as Row,
 } from "./AboutLayout";
-import styles from "./About.module.scss";
+import styles from "./About.module.css";
 
 export default function AboutCopyright() {
   return (
