+++ components/about/AboutHTML.tsx
@@ -6,7 +6,7 @@ import {
   AboutContainer as Container,
   AboutRow as Row,
 } from "./AboutLayout";
-import styles from "./About.module.scss";
+import styles from "./About.module.css";
 import { fetchAboutSectionFile } from "./about.helpers";
 
 interface Props {
