+++ components/distribution/Distribution.tsx
@@ -30,7 +30,7 @@ import { compareLocalized, formatInteger } from "@/i18n/format";
 import { DEFAULT_LOCALE, type SupportedLocale } from "@/i18n/locales";
 import { t } from "@/i18n/messages";
 import { fetchAllPages, fetchUrl } from "@/services/6529api";
-import styles from "./Distribution.module.scss";
+import styles from "./Distribution.module.css";
 
 interface Props {
   header: string;
