+++ components/gas-royalties/GasRoyalties.module.css
@@ -1,0 +1,181 @@
+.scrollContainer {
+  -ms-overflow-style: none;
+  overflow-x: auto;
+  scrollbar-width: none;
+}
+
+.scrollContainer::-webkit-scrollbar {
+  display: none;
+}
+
+.gasTable,
+.royaltiesTable {
+  width: 100%;
+}
+
+.gasTable tr td,
+.royaltiesTable tr td {
+  align-items: center;
+  vertical-align: middle;
+  white-space: nowrap;
+}
+
+.gasTable tr td span,
+.royaltiesTable tr td span {
+  float: left;
+  margin-top: 4px;
+  margin-bottom: 4px;
+}
+
+.gasTable tr td span a,
+.royaltiesTable tr td span a {
+  line-height: 40px;
+}
+
+.gasTable th,
+.royaltiesTable th {
+  line-height: 28px;
+  white-space: nowrap;
+}
+
+.gasTable th span,
+.royaltiesTable th span {
+  white-space: nowrap;
+}
+
+.gasTable td,
+.royaltiesTable td {
+  line-height: 40px;
+}
+
+.gasTable thead tr,
+.royaltiesTable thead tr {
+  background-color: transparent !important;
+}
+
+.gasTable tr:nth-child(odd),
+.royaltiesTable tr:nth-child(odd) {
+  background-color: rgb(30, 30, 30);
+}
+
+.gasTable > :not(caption) > * > *,
+.royaltiesTable > :not(caption) > * > * {
+  padding: 0.5rem;
+}
+
+.filterDropdown {
+  position: relative;
+  display: inline-block;
+  padding-right: 15px;
+  border: none;
+  background-color: transparent;
+  text-align: center;
+}
+
+.filterDropdown a,
+.filterDropdown .filterDropdownItem {
+  line-height: 30px;
+}
+
+.filterDropdown button {
+  cursor: pointer;
+  border: 0;
+  border-color: transparent !important;
+  background-color: transparent !important;
+  padding: 0;
+  color: white;
+  font-size: larger !important;
+  font-weight: bolder !important;
+}
+
+.filterDropdown button::after {
+  display: inline-block;
+  margin-top: 0.8rem;
+  margin-left: 0.255em;
+  border-top: 0.3em solid currentcolor;
+  border-right: 0.3em solid transparent;
+  border-left: 0.3em solid transparent;
+  vertical-align: 0.255em;
+  content: "";
+}
+
+@media only screen and (max-width: 800px) {
+  .filterDropdown button::after {
+    margin-top: 0.5rem;
+  }
+}
+
+.filterDropdownMenu {
+  position: absolute;
+  top: 100%;
+  left: 0;
+  z-index: 1000;
+  display: flex;
+  min-width: max-content;
+  flex-direction: column;
+  border: 1px solid rgb(51, 51, 51);
+  border-radius: 0.375rem;
+  background-color: rgb(34, 34, 34);
+  box-shadow: 0 0.5rem 1rem rgb(0 0 0 / 35%);
+  padding: 0.5rem 0;
+  text-align: left;
+}
+
+.filterDropdownItem {
+  padding: 0.25rem 1rem !important;
+  color: white;
+  font-size: 16px !important;
+  font-weight: normal !important;
+  text-align: left;
+  white-space: nowrap;
+}
+
+.filterDropdownItem::after {
+  display: none !important;
+}
+
+.filterDropdownItem:hover,
+.filterDropdownItem:focus {
+  background-color: rgb(51, 51, 51) !important;
+  color: white;
+}
+
+.filterDropdownDivider {
+  display: block;
+  margin: 0.5rem 0;
+  border-top: 1px solid rgb(51, 51, 51);
+}
+
+.filterDropdownHeader {
+  display: block;
+  padding: 0.25rem 1rem;
+  color: rgb(154, 154, 154);
+  font-size: 14px;
+  white-space: nowrap;
+}
+
+.collectionFocus {
+  border: 0;
+  background-color: transparent;
+  padding: 0;
+}
+
+.collectionFocus:hover {
+  color: white;
+  cursor: pointer;
+}
+
+.collectionFocusActive {
+  color: white !important;
+  transform: scale(1.02);
+}
+
+.infoIcon {
+  height: 18px;
+  cursor: pointer;
+  color: rgb(154, 154, 154);
+}
+
+.infoIcon:hover {
+  color: white;
+}
