--- components/gas-royalties/GasRoyalties.module.scss
@@ -1,167 +1,0 @@
-@use "../../styles/variables.scss";
-
-.scrollContainer {
-  -ms-overflow-style: none; /* Internet Explorer 10+ */
-  scrollbar-width: none; /* Firefox */
-  overflow-x: auto;
-}
-
-.scrollContainer::-webkit-scrollbar {
-  display: none;
-}
-
-.gasTable,
-.royaltiesTable {
-  width: 100%;
-
-  tr td {
-    white-space: nowrap;
-    align-items: center;
-    vertical-align: middle;
-    span {
-      float: left;
-      margin-top: 4px;
-      margin-bottom: 4px;
-      a {
-        line-height: 40px;
-      }
-    }
-  }
-
-  th {
-    white-space: nowrap;
-    line-height: 28px;
-  }
-
-  th span {
-    white-space: nowrap;
-  }
-
-  td {
-    line-height: 40px;
-  }
-
-  thead tr {
-    background-color: transparent !important;
-  }
-
-  tr:nth-child(odd) {
-    background-color: variables.$almost-very-dark-grey;
-  }
-}
-
-.gasTable > :not(caption) > * > *,
-.royaltiesTable > :not(caption) > * > * {
-  padding: 0.5rem;
-}
-
-.filterDropdown {
-  display: inline-block;
-  position: relative;
-  text-align: center;
-  background-color: transparent;
-  border: none;
-  padding-right: 15px;
-  a,
-  .filterDropdownItem {
-    line-height: 30px;
-  }
-  button {
-    font-size: larger !important;
-    font-weight: bolder !important;
-    cursor: pointer;
-    padding: 0;
-    background-color: transparent !important;
-    border: 0;
-    border-color: transparent !important;
-    color: variables.$font-color;
-    &::after {
-      border-left: 0.3em solid transparent;
-      border-right: 0.3em solid transparent;
-      border-top: 0.3em solid currentcolor;
-      content: "";
-      display: inline-block;
-      margin-left: 0.255em;
-      margin-top: 0.8rem;
-      vertical-align: 0.255em;
-      @media only screen and (max-width: 800px) {
-        margin-top: 0.5rem;
-      }
-    }
-  }
-}
-
-.filterDropdownMenu {
-  background-color: variables.$very-dark-grey;
-  border: 1px solid variables.$dark-grey;
-  border-radius: 0.375rem;
-  box-shadow: 0 0.5rem 1rem rgb(0 0 0 / 35%);
-  display: flex;
-  flex-direction: column;
-  left: 0;
-  min-width: max-content;
-  padding: 0.5rem 0;
-  position: absolute;
-  text-align: left;
-  top: 100%;
-  z-index: 1000;
-}
-
-.filterDropdownItem {
-  color: variables.$font-color;
-  font-size: variables.$font-size !important;
-  font-weight: normal !important;
-  padding: 0.25rem 1rem !important;
-  text-align: left;
-  white-space: nowrap;
-
-  &::after {
-    display: none !important;
-  }
-
-  &:hover,
-  &:focus {
-    background-color: variables.$dark-grey !important;
-    color: white;
-  }
-}
-
-.filterDropdownDivider {
-  border-top: 1px solid variables.$dark-grey;
-  display: block;
-  margin: 0.5rem 0;
-}
-
-.filterDropdownHeader {
-  color: variables.$lightest-grey;
-  display: block;
-  font-size: variables.$font-smaller;
-  padding: 0.25rem 1rem;
-  white-space: nowrap;
-}
-
-.collectionFocus {
-  background-color: transparent;
-  border: 0;
-  padding: 0;
-
-  &:hover {
-    color: variables.$font-color;
-    cursor: pointer;
-  }
-}
-
-.collectionFocusActive {
-  color: variables.$font-color !important;
-  transform: scale(1.02);
-}
-
-.infoIcon {
-  color: variables.$font-color-h;
-  height: 18px;
-  cursor: pointer;
-
-  &:hover {
-    color: variables.$font-color;
-  }
-}
