+++ components/mapping-tools/ConsolidationMappingTool.tsx
@@ -9,7 +9,7 @@ import {
   MappingToolSubmitButton,
   MappingToolUpload,
 } from "./MappingToolControls";
-import styles from "./MappingTool.module.scss";
+import styles from "./MappingTool.module.css";
 
 const csvParser = require("csv-parser");
