--- components/mapping-tools/MappingTool.module.scss
@@ -1,97 +1,0 @@
-@use "sass:color";
-@use "../../styles/variables.scss" as variables;
-
-.placeholder {
-  height: 500px;
-}
-
-.formInput {
-  color: variables.$dark-grey !important;
-  border-radius: 0 !important;
-  border-color: variables.$lightest-grey;
-  cursor: pointer;
-
-  &:focus {
-    border-color: variables.$project-blue !important;
-  }
-}
-
-.formInputHidden {
-  display: none;
-}
-
-.submitBtn {
-  width: 100%;
-  border-radius: 0;
-  background-color: variables.$project-green;
-  padding-left: 15px;
-  padding-right: 25px;
-  border: none;
-  padding: 12px 20px;
-  cursor: pointer;
-  margin-right: 20px;
-  margin-top: 15px;
-  color: white;
-
-  &:active {
-    background-color: variables.$project-green !important;
-  }
-
-  &:hover {
-    background-color: color.scale(variables.$project-green, $lightness: -5%);
-  }
-}
-
-.submitBtnDisabled {
-  pointer-events: none !important;
-  background-color: color.scale(variables.$project-green, $lightness: -10%);
-  color: variables.$font-color-h;
-}
-
-.loader {
-  width: 20px;
-  height: 20px;
-  margin-left: 5px;
-}
-
-.toolArea {
-  padding: 25px 30px;
-  background-color: variables.$almost-very-dark-grey;
-}
-
-.uploadArea {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 100%;
-  padding: 50px;
-  background-color: variables.$very-dark-grey;
-  align-items: center;
-  gap: 2rem;
-  border: 2px dashed variables.$lighter-grey;
-  color: inherit;
-  cursor: pointer;
-  font: inherit;
-  text-align: center;
-
-  &:hover {
-    border: 2px dashed variables.$lightest-grey;
-  }
-}
-
-.uploadAreaActive {
-  border: 2px dashed variables.$lightest-grey;
-}
-
-.uploadIcon {
-  height: 4rem;
-}
-
-.clearBtn {
-  height: 25px;
-  cursor: pointer;
-
-  &:focus {
-    outline: none !important;
-  }
-}
