--- components/timeline/Timeline.module.scss
@@ -1,177 +1,0 @@
-@use "../../styles/variables.scss";
-
-.timeline {
-  position: relative;
-  margin: 0 auto;
-  word-wrap: break-word;
-
-  h4,
-  h5,
-  p {
-    color: variables.$very-dark-grey !important;
-  }
-
-  a {
-    color: variables.$very-dark-grey !important;
-
-    &:hover {
-      opacity: 0.8;
-    }
-  }
-}
-
-.timeline::after {
-  content: "";
-  position: absolute;
-  width: 6px;
-  background-color: white;
-  top: 0;
-  bottom: 0;
-  left: 50%;
-  margin-left: -3px;
-}
-
-.timelineContainer {
-  padding: 20px 26px;
-  position: relative;
-  background-color: inherit;
-  width: 50%;
-}
-
-.timelineContainer::after {
-  content: "";
-  position: absolute;
-  width: 25px;
-  height: 25px;
-  right: -13px;
-  background-color: white;
-  border: 6px solid variables.$salmon;
-  top: 25px;
-  border-radius: 50%;
-  z-index: 1;
-}
-
-.left,
-.right {
-  left: 0;
-}
-
-.left::before {
-  content: " ";
-  height: 0;
-  position: absolute;
-  top: 28px;
-  width: 0;
-  z-index: 1;
-  right: 16px;
-  border: medium solid white;
-  border-width: 10px 0 10px 10px;
-  border-color: transparent transparent transparent white;
-}
-
-.right::before {
-  content: " ";
-  height: 0;
-  position: absolute;
-  top: 28px;
-  width: 0;
-  z-index: 1;
-  left: 16px;
-  border: medium solid white;
-  border-width: 10px 10px 10px 0;
-  border-color: transparent white transparent transparent;
-}
-
-.content {
-  color: variables.$very-dark-grey;
-  padding: 20px;
-  background-color: white;
-  position: relative;
-  border-radius: 6px;
-}
-
-.timeline::after {
-  left: 31px;
-}
-
-.timelineContainer {
-  width: 100%;
-  padding-left: 70px;
-  padding-right: 25px;
-}
-
-.timelineContainer::before {
-  left: 60px;
-  border: medium solid white;
-  border-width: 10px 10px 10px 0;
-  border-color: transparent white transparent transparent;
-}
-
-.left::after,
-.right::after {
-  left: 18px;
-}
-
-.externalLink {
-  text-decoration: none;
-}
-
-.linkIcon {
-  height: 25px;
-  width: 25px;
-  cursor: pointer;
-}
-
-.timelineTable {
-  tr {
-    border-bottom: 1px solid #333 !important;
-  }
-  tr td {
-    white-space: nowrap;
-  }
-  tr td:first-child {
-    width: 0px;
-  }
-  tr th {
-    min-width: 80px;
-    white-space: nowrap;
-    line-height: 30px;
-  }
-}
-
-.timelineTableScrollContainer {
-  -ms-overflow-style: none; /* Internet Explorer 10+ */
-  scrollbar-width: none; /* Firefox */
-  overflow-x: auto;
-}
-
-.timelineTableScrollContainer::-webkit-scrollbar {
-  display: none;
-}
-
-.timelineTable > :not(caption) > * > * {
-  padding: 1.5rem !important;
-}
-
-.timelineMediaImage {
-  height: auto;
-  width: 100%;
-  min-width: 150px;
-}
-
-.changesUl {
-  padding-left: 1.5rem;
-  padding-right: 1.5rem;
-}
-
-.changeLink {
-  margin-top: 0;
-  margin-bottom: 0;
-  font-size: smaller;
-  overflow-wrap: anywhere;
-}
-
-.metadataValue {
-  overflow-wrap: anywhere;
-  white-space: pre-line;
-}
